Visualisation: extract drawing into a function
Currently, the d3js data file was hardcoded in the javascript file.
It is a computation of the current date, in yy-mm-dd format.
This does not allows the user to view previous dates, for instance.

This patch extracts the code which performs the graph drawing.
It is parametrized by the filename, which is now(). at the moment

Signed-off-by: Mattijs Korpershoek <mattijs.korpershoek@gmail.com>